refactor: change name of /persistance to /persistence
