+++ src/main/java/it/smartcommunitylab/aac/config/DatabaseConfig.java
@@ -59,7 +59,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
         "it.smartcommunitylab.aac.oauth.persistence",
         "it.smartcommunitylab.aac.oidc.persistence",
         "it.smartcommunitylab.aac.password.persistence",
-        "it.smartcommunitylab.aac.otp.persistance",
+        "it.smartcommunitylab.aac.otp.persistence",
         "it.smartcommunitylab.aac.realms.persistence",
         "it.smartcommunitylab.aac.roles.persistence",
         "it.smartcommunitylab.aac.saml.persistence",
@@ -82,7 +82,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
         "it.smartcommunitylab.aac.oauth.persistence",
         "it.smartcommunitylab.aac.oidc.persistence",
         "it.smartcommunitylab.aac.password.persistence",
-        "it.smartcommunitylab.aac.otp.persistance",
+        "it.smartcommunitylab.aac.otp.persistence",
         "it.smartcommunitylab.aac.realms.persistence",
         "it.smartcommunitylab.aac.roles.persistence",
         "it.smartcommunitylab.aac.saml.persistence",
@@ -169,7 +169,7 @@ public class DatabaseConfig {
             "it.smartcommunitylab.aac.oauth.persistence",
             "it.smartcommunitylab.aac.oidc.persistence",
             "it.smartcommunitylab.aac.password.persistence",
-            "it.smartcommunitylab.aac.otp.persistance",
+            "it.smartcommunitylab.aac.otp.persistence",
             "it.smartcommunitylab.aac.realms.persistence",
             "it.smartcommunitylab.aac.roles.persistence",
             "it.smartcommunitylab.aac.saml.persistence",
