+++ src/main/java/it/smartcommunitylab/aac/otp/provider/OtpCleanupScheduler.java
@@ -1,7 +1,7 @@
 package it.smartcommunitylab.aac.otp.provider;
 
-import it.smartcommunitylab.aac.otp.persistance.InternalUserOtpEntity;
-import it.smartcommunitylab.aac.otp.persistance.InternalUserOtpEntityRepository;
+import it.smartcommunitylab.aac.otp.persistence.InternalUserOtpEntity;
+import it.smartcommunitylab.aac.otp.persistence.InternalUserOtpEntityRepository;
 import java.util.List;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
