+++ src/main/java/it/smartcommunitylab/aac/otp/service/OtpUserCredentialsService.java
@@ -5,8 +5,8 @@ import it.smartcommunitylab.aac.common.NoSuchCredentialException;
 import it.smartcommunitylab.aac.common.RegistrationException;
 import it.smartcommunitylab.aac.credentials.persistence.UserCredentialsService;
 import it.smartcommunitylab.aac.otp.model.InternalUserOtp;
-import it.smartcommunitylab.aac.otp.persistance.InternalUserOtpEntity;
-import it.smartcommunitylab.aac.otp.persistance.InternalUserOtpEntityRepository;
+import it.smartcommunitylab.aac.otp.persistence.InternalUserOtpEntity;
+import it.smartcommunitylab.aac.otp.persistence.InternalUserOtpEntityRepository;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
